+++ diesel/src/internal/derives.rs
@@ -31,7 +31,7 @@ pub mod multiconnection {
         BoxedLimitOffsetClause, LimitOffsetClause,
     };
     #[doc(hidden)]
-    pub use crate::query_builder::returning_clause::ReturningClause;
+    pub use crate::query_builder::returning::returning_clause::ReturningClause;
     #[doc(hidden)]
     pub use crate::query_builder::select_statement::SelectStatement;
     #[doc(hidden)]