+++ diesel/src/query_builder/delete_statement/mod.rs
@@ -1,7 +1,8 @@
 use crate::backend::DieselReserveSpecialization;
 use crate::dsl::{Filter, IntoBoxed, OrFilter};
 use crate::expression::{AppearsOnTable, SelectableExpression};
-use crate::query_builder::returning_clause::*;
+use crate::query_builder::returning::returning_clause::*;
+use crate::query_builder::returning::returning_expression::{self, ReturningExpression};
 use crate::query_builder::where_clause::*;
 use crate::query_builder::*;
 use crate::query_dsl::RunQueryDsl;
@@ -269,7 +270,7 @@ where
 impl<T, U> AsQuery for DeleteStatement<T, U, NoReturningClause>
 where
     T: Table,
-    T::AllColumns: SelectableExpression<T> + ReturningExpression<DeleteStmt, T>,
+    T::AllColumns: SelectableExpression<T> + ReturningExpression<returning_expression::DeleteStmt, T>,
     DeleteStatement<T, U, ReturningClause<T::AllColumns>>: Query,
 {
     type SqlType = <Self::Query as Query>::SqlType;
@@ -283,9 +284,9 @@ where
 impl<T, U, Ret> Query for DeleteStatement<T, U, ReturningClause<Ret>>
 where
     T: Table,
-    Ret: ReturningExpression<DeleteStmt, T>,
+    Ret: ReturningExpression<returning_expression::DeleteStmt, T>,
 {
-    type SqlType = <Ret as ReturningExpression<DeleteStmt, T>>::SqlType;
+    type SqlType = <Ret as ReturningExpression<returning_expression::DeleteStmt, T>>::SqlType;
 }
 
 impl<T, U, Ret, Conn> RunQueryDsl<Conn> for DeleteStatement<T, U, Ret> where T: QuerySource {}