+++ diesel/src/query_builder/insert_statement/mod.rs
@@ -11,7 +11,8 @@ pub(crate) use self::private::Insert;
 )]
 pub(crate) use self::private::{InsertOrIgnore, Replace};
 
-use super::returning_clause::*;
+use super::returning::returning_clause::*;
+use super::returning::returning_expression::{InsertStmtKind, ReturningExpression};
 use crate::backend::{DieselReserveSpecialization, SqlDialect, sql_dialect};
 use crate::expression::grouped::Grouped;
 use crate::expression::operators::Eq;
@@ -285,8 +286,9 @@ where
 impl<T, U, Op> AsQuery for InsertStatement<T, U, Op, NoReturningClause>
 where
     T: Table,
+    U: InsertStmtKind,
     InsertStatement<T, U, Op, ReturningClause<T::AllColumns>>: Query,
-    T::AllColumns: ReturningExpression<InsertStmt, T>,
+    T::AllColumns: ReturningExpression<U::StmtKind, T>,
 {
     type SqlType = <Self::Query as Query>::SqlType;
     type Query = InsertStatement<T, U, Op, ReturningClause<T::AllColumns>>;
@@ -299,9 +301,10 @@ where
 impl<T, U, Op, Ret> Query for InsertStatement<T, U, Op, ReturningClause<Ret>>
 where
     T: QuerySource,
-    Ret: ReturningExpression<InsertStmt, T> + NonAggregate,
+    U: InsertStmtKind,
+    Ret: ReturningExpression<U::StmtKind, T> + NonAggregate,
 {
-    type SqlType = <Ret as ReturningExpression<InsertStmt, T>>::SqlType;
+    type SqlType = <Ret as ReturningExpression<U::StmtKind, T>>::SqlType;
 }
 
 impl<T: QuerySource, U, Op, Ret, Conn> RunQueryDsl<Conn> for InsertStatement<T, U, Op, Ret> {}
@@ -594,7 +597,7 @@ mod private {
         type Table = T;
         type Op = Op;
         type Values = ();
-        type Ret = crate::query_builder::returning_clause::NoReturningClause;
+        type Ret = crate::query_builder::returning::returning_clause::NoReturningClause;
     }
 
     impl<T, U, Op, Ret> InsertAutoTypeHelper for InsertStatement<T, U, Op, Ret>